requirements.txt
Flask==2.3.2
Flask-CORS==3.1.1
easyocr==1.4.1
numpy==1.24.0
Pillow==9.5.0
PyMuPDF==1.23.5